Used built it Preferences to create settings. Removed unnecessary code and binded customization activity
