+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    package="com.yanrou.dawnisland">
+  xmlns:tools="http://schemas.android.com/tools"
+  package="com.yanrou.dawnisland">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -9,24 +9,32 @@
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
 
     <application
-        android:name=".MyApplication"
-        android:allowBackup="true"
-        android:hardwareAccelerated="true"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
-        android:networkSecurityConfig="@xml/network_security_config"
-        android:requestLegacyExternalStorage="true"
-        android:roundIcon="@mipmap/ic_launcher_round"
-        android:supportsRtl="true"
-        android:theme="@style/MyAppTheme"
-        tools:ignore="GoogleAppIndexingWarning">
+      android:name=".MyApplication"
+      android:allowBackup="true"
+      android:hardwareAccelerated="true"
+      android:icon="@mipmap/ic_launcher"
+      android:label="@string/app_name"
+      android:networkSecurityConfig="@xml/network_security_config"
+      android:requestLegacyExternalStorage="true"
+      android:roundIcon="@mipmap/ic_launcher_round"
+      android:supportsRtl="true"
+      android:theme="@style/MyAppTheme"
+      tools:ignore="GoogleAppIndexingWarning">
+        <activity
+          android:name=".settings.SettingsActivity"
+          android:label="@string/title_activity_settings" />
         <activity android:name=".imageviewer.ImageViewerView" />
-        <activity android:name=".settings.SettingsActivity" />
+        <activity android:name=".settings.CustomizeSizesActivity">
+            <intent-filter>
+                <action android:name="com.yanrou.dawnisland.settings.CustomizeSizesActivity" />
+                <category android:name="android.intent.category.DEFAULT" />
+            </intent-filter>
+        </activity>
         <activity android:name=".CookiesManageActivity" />
         <activity
-            android:name=".content.SeriesContentActivity"
-            android:launchMode="singleTop"
-            android:windowSoftInputMode="adjustNothing" />
+          android:name=".content.SeriesContentActivity"
+          android:launchMode="singleTop"
+          android:windowSoftInputMode="adjustNothing" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
