+++ app/src/main/java/com/yanrou/dawnisland/MainActivity.java
@@ -86,7 +86,6 @@ public class MainActivity extends AppCompatActivity {
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
         switch (item.getItemId()) {
             case R.id.setting:
-                //Intent intent = new Intent(MainActivity.this, SettingsActivity.class);
                 Intent intent = new Intent(MainActivity.this, SettingsActivity.class);
                 startActivity(intent);
                 break;
