+++ app/src/main/res/layout/activity_customize_sizes.xml
@@ -1,0 +1,18 @@
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <ScrollView
+      android:id="@+id/customize_sizes"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" >
+        <LinearLayout
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+            android:orientation="vertical">
+
+        </LinearLayout>
+
+
+    </ScrollView>
+</LinearLayout>
