+++ app/src/main/res/layout/settings_activity.xml
@@ -1,0 +1,9 @@
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+  android:layout_width="match_parent"
+  android:layout_height="match_parent">
+
+    <FrameLayout
+      android:id="@+id/settings"
+      android:layout_width="match_parent"
+      android:layout_height="match_parent" />
+</LinearLayout>
